# packages/api/openapi.private.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Prontiq Private Account API",
        "version": "1.0.0",
        "description": "Private Clerk-authenticated account and console API contracts."
    },
    "components": {
        "securitySchemes": {
            "ClerkJwt": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Clerk session token; sent as `Authorization: Bearer <jwt>`."
            }
        },
...
    },
    "paths": {
        "/v1/account/setup": {
            "post": {
                "summary": "Recover or initialise an org envelope (Clerk JWT auth)",
                "description": "Idempotent. Resolves the caller's verified primary email via the Clerk Backend API and provisions the org envelope. The Clerk org id is the active Prontiq and Lago customer identity; Stripe remains only Lago's payment rail.",
                "security": [
                    {
                        "ClerkJwt": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Org envelope already provisioned (replay-safe)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "created",
                                                "already_exists"
                                            ]
                                        },
                                        "orgId": {
                                            "type": "string",
                                            "description": "Clerk organization id and active Prontiq/Lago customer identity."
                                        },
                                        "emailSent": {
                                            "type": "boolean",
                                            "description": "True only when the welcome email was best-effort sent for a freshly-created envelope. Absent on already_exists replays."
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "orgId"
                                    ]
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Org envelope freshly created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "created",
                                                "already_exists"
                                            ]
                                        },
                                        "orgId": {
                                            "type": "string",
                                            "description": "Clerk organization id and active Prontiq/Lago customer identity."
                                        },
                                        "emailSent": {
                                            "type": "boolean",
                                            "description": "True only when the welcome email was best-effort sent for a freshly-created envelope. Absent on already_exists replays."
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "orgId"
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "JWT missing org_id claim (NO_ACTIVE_ORG)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "string"
                                                },
                                                "message": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "integer"
                                                },
                                                "request_id": {
                                                    "type": "string"
                                                },
                                                "details": {
                                                    "type": "object",
                                                    "additionalProperties": {}
                                                }
                                            },
                                            "required": [
                                                "code",
                                                "message",
                                                "status",
                                                "request_id"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing/invalid/expired JWT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "string"
                                                },
                                                "message": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "integer"
                                                },
                                                "request_id": {
                                                    "type": "string"
                                                },
                                                "details": {
                                                    "type": "object",
                                                    "additionalProperties": {}
                                                }
                                            },
                                            "required": [
                                                "code",
                                                "message",
                                                "status",
                                                "request_id"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Fatal failure — operator-facing fix required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "string"
                                                },
                                                "message": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "integer"
                                                },
                                                "request_id": {
                                                    "type": "string"
                                                },
                                                "details": {
                                                    "type": "object",
                                                    "additionalProperties": {}
                                                }
                                            },
                                            "required": [
                                                "code",
                                                "message",
                                                "status",
                                                "request_id"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Retryable failure — Lago/Clerk/DDB transient",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "string"
                                                },
                                                "message": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "integer"
                                                },
                                                "request_id": {
                                                    "type": "string"
                                                },
                                                "details": {
                                                    "type": "object",
                                                    "additionalProperties": {}
                                                }
                                            },
                                            "required": [
                                                "code",
                                                "message",
                                                "status",
                                                "request_id"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/account/keys": {
            "get": {
                "summary": "List active keys for the caller's org (member-allowed)",
                "description": "Returns active keys with masked prefix only. The org envelope is excluded by sentinel filter; revoked keys are excluded by the active flag.",
                "security": [
                    {
                        "ClerkJwt": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Active keys for this org",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "keys": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "keyId": {
                                                        "type": "string"
                                                    },
                                                    "keyPrefix": {
                                                        "type": "string"
                                                    },
                                                    "label": {
                                                        "type": "string"
                                                    },
                                                    "createdAt": {
                                                        "type": "string"
                                                    },
                                                    "lastUsedAt": {
                                                        "type": [
                                                            "string",
                                                            "null"
                                                        ]
                                                    },
                                                    "active": {
                                                        "type": "boolean"
                                                    },
                                                    "products": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        }
                                                    }
                                                },
                                                "required": [
                                                    "keyId",
                                                    "keyPrefix",
                                                    "createdAt",
                                                    "lastUsedAt",
                                                    "active",
                                                    "products"
                                                ]
                                            }
                                        }
                                    },
                                    "required": [
                                        "keys"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing/invalid JWT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "string"
                                                },
                                                "message": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "integer"
                                                },
                                                "request_id": {
                                                    "type": "string"
                                                },
                                                "details": {
                                                    "type": "object",
                                                    "additionalProperties": {}
                                                }
                                            },
                                            "required": [
                                                "code",
                                                "message",
                                                "status",
                                                "request_id"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "string"
                                                },
                                                "message": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "integer"
                                                },
                                                "request_id": {
                                                    "type": "string"
                                                },
                                                "details": {
                                                    "type": "object",
                                                    "additionalProperties": {}
                                                }
                                            },
                                            "required": [
                                                "code",
                                                "message",
                                                "status",
                                                "request_id"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}